// api/generated/models/UpdatePartnerAgreementDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdatePartnerAgreementDto = {
    /**
     * 이용약관 동의
     */
    termsOfUse?: boolean;
    /**
     * 개인정보 처리방침 동의
     */
    privacyPolicy?: boolean;
    /**
     * SMS 마케팅 동의
     */
    marketingSms?: boolean;
    /**
     * 이메일 마케팅 동의
     */
    marketingEmail?: boolean;
};